requirements_unified.txt
# Requirements Unificados para Proyecto_J
# ======================================

# Core Data Science
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Streamlit and Web
streamlit>=1.28.0
streamlit-option-menu>=0.3.6

# Statistical Analysis
statsmodels>=0.14.0
scikit-learn>=1.3.0

# Data Loading
pyreadstat>=1.2.0
openpyxl>=3.1.0
xlrd>=2.0.0

# Time Series Analysis
statsmodels>=0.14.0

# Machine Learning
scikit-learn>=1.3.0

# Data Processing
openpyxl>=3.1.0
xlrd>=2.0.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0

# Additional Analysis Tools
pingouin>=0.5.0
scipy>=1.10.0

# File Handling
pathlib2>=2.3.0

# Configuration
pyyaml>=6.0.0
toml>=0.10.0

# Logging and Monitoring
structlog>=23.1.0

# Performance
numba>=0.57.0

# Optional: Advanced Features
# Uncomment if needed
# tensorflow>=2.13.0
# torch>=2.0.0
# transformers>=4.30.0 